% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whibo_clustering.R
\name{plot.whibo_cluster}
\alias{plot.whibo_cluster}
\title{Plot WhiBo Cluster Representatives}
\usage{
\method{plot}{whibo_cluster}(x, ...)
}
\arguments{
\item{x}{WhiBo Cluster model.}

\item{...}{None of those will be used.}
}
\value{
Line plot with Cluster representatives
}
\description{
Plot WhiBo Cluster Representatives
}
\examples{
data <- iris[, 1:4] #Numerical data only

model <- whibo_clustering(data = data, k = 3)
plot(model)
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
