% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whibo_clustering.R
\name{print.whibo_cluster}
\alias{print.whibo_cluster}
\title{Show White-Box Cluster Algorithm model}
\usage{
\method{print}{whibo_cluster}(x, ...)
}
\arguments{
\item{x}{WhiBo Cluster model.}

\item{...}{None of those will be used.}
}
\value{
Summary text about Cluster model.
}
\description{
Show White-Box Cluster Algorithm model
}
\examples{
data <- iris[, 1:4] #Numerical data only

model <- whibo_clustering(data = data, k = 3)
print(model)
}
\seealso{
\code{show.whibo_cluster}, \code{head.whibo_cluster}, \code{summary.whibo_cluster}
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
