% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whibo_clustering.R
\name{show.whibo_cluster}
\alias{show.whibo_cluster}
\title{Show White-Box Cluster Algorithm model}
\usage{
show.whibo_cluster(object)
}
\arguments{
\item{object}{WhiBo Cluster model.}
}
\value{
Summary text about Cluster model.
}
\description{
Show White-Box Cluster Algorithm model
}
\examples{
data <- iris[, 1:4] #Numerical data only

model <- whibo_clustering(data = data, k = 3)
show(model)
}
\seealso{
\code{print.whibo_cluster}, \code{head.whibo_cluster}, \code{summary.whibo_cluster}
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
