% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whibo_clustering.R
\name{summary.whibo_cluster}
\alias{summary.whibo_cluster}
\title{Show White-Box Cluster Algorithm model}
\usage{
\method{summary}{whibo_cluster}(object, ...)
}
\arguments{
\item{object}{WhiBo Cluster model.}

\item{...}{None of those will be used.}
}
\value{
Summary text about Cluster model.
}
\description{
Show White-Box Cluster Algorithm model
}
\examples{
data <- iris[, 1:4] #Numerical data only

model <- whibo_clustering(data = data, k = 3)
summary(model)

}
\seealso{
\code{print.whibo_cluster}
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
