% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_assignment.R
\name{wc_assign_fidelity_numerical}
\alias{wc_assign_fidelity_numerical}
\title{Assign data points using Fidelity (numerical version) distance.}
\usage{
wc_assign_fidelity_numerical(data, centroids)
}
\arguments{
\item{data}{A dataset for which data points needs to be assigned to Cluster Representatives.}

\item{centroids}{Cluster representatives.}
}
\value{
A vector of assignments.
}
\description{
Assign data points using Fidelity (numerical version) distance.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
