% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_assignment.R
\docType{data}
\name{wc_assign_types}
\alias{wc_assign_types}
\title{Data frame for possible values of assignment types.}
\format{An object of class \code{data.frame} with 18 rows and 2 columns.}
\usage{
wc_assign_types
}
\description{
Data frame for possible values of assignment types.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
\keyword{datasets}
