% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_assignment.R
\name{wc_assign_whittaker}
\alias{wc_assign_whittaker}
\title{Assign data points using Whittaker distance.}
\usage{
wc_assign_whittaker(data, centroids)
}
\arguments{
\item{data}{A dataset for which data points needs to be assigned to Cluster Representatives.}

\item{centroids}{Cluster representatives.}
}
\value{
A vector of assignments.
}
\description{
Assign data points using Whittaker distance.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
