% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_assignment.R
\name{wc_assignment}
\alias{wc_assignment}
\title{General Component for Assignment of data points to Cluster Representatives.}
\usage{
wc_assignment(data, centroids, assignment_type)
}
\arguments{
\item{data}{A dataset for which data points needs to be assigned to Cluster Representatives.}

\item{centroids}{Cluster representatives.}

\item{assignment_type}{String which signal which assignment type to be used. Check \code{wc_assign_types} for possible values.}
}
\value{
A vector of assignments.
}
\description{
General Component for Assignment of data points to Cluster Representatives.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
