% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_cluster_performance.R
\name{wc_eval_between_sum_of_squares}
\alias{wc_eval_between_sum_of_squares}
\title{Calculate between (Clusters) sum of squares}
\usage{
wc_eval_between_sum_of_squares(data, centroids, assignment)
}
\arguments{
\item{data}{A dataset for which between sum of squared should be calculated.}

\item{centroids}{A data frame of cluster representatives.}

\item{assignment}{Vector of assignments.}
}
\value{
A vector of number which shows between (clusters) sum of squares.
}
\description{
Calculate between (Clusters) sum of squares
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
