% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_cluster_performance.R
\name{wc_eval_g_plus}
\alias{wc_eval_g_plus}
\title{Calculate G+ index internal Cluster evaluation measure}
\usage{
wc_eval_g_plus(data, assignment)
}
\arguments{
\item{data}{A dataset for which internal cluster quality should be calculated.}

\item{assignment}{Vector of assignments.}
}
\value{
A value of internal cluster quality evaluation measure.
}
\description{
Calculate G+ index internal Cluster evaluation measure
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
