% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_cluster_performance.R
\name{wc_eval_total_sum_of_squares}
\alias{wc_eval_total_sum_of_squares}
\title{Calculate total sum of squares}
\usage{
wc_eval_total_sum_of_squares(data)
}
\arguments{
\item{data}{A dataset for which total sum of squared should be calculated.}
}
\value{
A number which shows total sum of squares.
}
\description{
Calculate total sum of squares
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
