% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_initialization.R
\docType{data}
\name{wc_init_types}
\alias{wc_init_types}
\title{Data frame for possible values of initialization types.}
\format{An object of class \code{data.frame} with 10 rows and 2 columns.}
\usage{
wc_init_types
}
\description{
Data frame for possible values of initialization types.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
\keyword{datasets}
