% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_initialization.R
\name{wc_initialize}
\alias{wc_initialize}
\title{General Component for Initialization of Cluster Representatives.}
\usage{
wc_initialize(data, k = 3, initialization_type)
}
\arguments{
\item{data}{A dataset for which Cluster Representatives needs to be initialized.}

\item{k}{A number of Cluster Representatives.}

\item{initialization_type}{String which signal which initialization type to be used. Check \code{wc_init_types} for possible values.}
}
\value{
As a result initial Cluster Representatives are obtained. Result is in for of data.frame or data.matrix.
}
\description{
General Component for Initialization of Cluster Representatives.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
