% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_normalization.R
\name{wc_norm_mean}
\alias{wc_norm_mean}
\title{Mean (Max-Min like) normalization of data.}
\usage{
wc_norm_mean(data, model = NULL)
}
\arguments{
\item{data}{A dataset which needs to be normalized.}

\item{model}{Additional data needed for future data to be normalized using same normalization technique.}
}
\value{
As a result normalized data are obtained. Result is in for of data.frame or data.matrix.
}
\description{
Mean (Max-Min like) normalization of data.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
