% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_normalization.R
\docType{data}
\name{wc_norm_types}
\alias{wc_norm_types}
\title{Data frame for possible values of normalization types.}
\format{An object of class \code{data.frame} with 13 rows and 2 columns.}
\usage{
wc_norm_types
}
\description{
Data frame for possible values of normalization types.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
\keyword{datasets}
