% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_normalization.R
\name{wc_normalize}
\alias{wc_normalize}
\title{General (Optional) Component for Normalization of data.}
\usage{
wc_normalize(data, normalization_type)
}
\arguments{
\item{data}{A dataset which needs to be normalized.}

\item{normalization_type}{String which signal which normalization type to be used. Check \code{wc_norm_types} for possible values.}
}
\value{
As a result normalized data are obtained. Result is in for of data.frame or data.matrix.
}
\description{
General (Optional) Component for Normalization of data.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
