% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_recalculate.R
\name{wc_recalc_geometric_mean}
\alias{wc_recalc_geometric_mean}
\title{Solution for Cluster Representative update which uses Geometric mean.}
\usage{
wc_recalc_geometric_mean(data, assignment)
}
\arguments{
\item{data}{A dataset for which Cluster Representatives needs to be updated.}

\item{assignment}{Vector of Cluster assignments.}
}
\value{
As a result new Cluster Representatives are obtained. Result is in for of data.frame or data.matrix.
}
\description{
Solution for Cluster Representative update which uses Geometric mean.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
