% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_recalculate.R
\name{wc_recalc_mean}
\alias{wc_recalc_mean}
\title{Solution for Cluster Representative update which uses Mean.}
\usage{
wc_recalc_mean(data, assignment, old_centroids = NULL)
}
\arguments{
\item{data}{A dataset for which Cluster Representatives needs to be updated.}

\item{assignment}{Vector of Cluster assignments.}

\item{old_centroids}{Old centroids.}
}
\value{
As a result new Cluster Representatives are obtained. Result is in for of data.frame or data.matrix.
}
\description{
Solution for Cluster Representative update which uses Mean.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
