% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_recalculate.R
\name{wc_recalc_online_midrange}
\alias{wc_recalc_online_midrange}
\title{Solution for Cluster Representative update which uses Online Midrange.}
\usage{
wc_recalc_online_midrange(data, assignment, assignment_type)
}
\arguments{
\item{data}{A dataset for which Cluster Representatives needs to be updated.}

\item{assignment}{Vector of Cluster assignments.}

\item{assignment_type}{Assignment type to be used.}
}
\value{
As a result new Cluster Representatives are obtained. Result is in for of data.frame or data.matrix.
}
\description{
Solution for Cluster Representative update which uses Online Midrange.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
