% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_recalculate.R
\name{wc_recalculate}
\alias{wc_recalculate}
\title{General Component for Cluster Representative update.}
\usage{
wc_recalculate(data, assignment, recalculate_type, assignment_type = NULL)
}
\arguments{
\item{data}{A dataset for which Cluster Representatives needs to be updated.}

\item{assignment}{Vector of Cluster assignments.}

\item{recalculate_type}{String which signal which update type to be used. Check \code{wc_recalculate_types} for possible values.}
}
\value{
As a result new Cluster Representatives are obtained. Result is in for of data.frame or data.matrix.
}
\description{
General Component for Cluster Representative update.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
