% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wc_recalculate.R
\docType{data}
\name{wc_recalculate_types}
\alias{wc_recalculate_types}
\title{Data frame for possible values of recalculate types.}
\format{An object of class \code{data.frame} with 18 rows and 2 columns.}
\usage{
wc_recalculate_types
}
\description{
Data frame for possible values of recalculate types.
}
\author{
Sandro Radovanovic \email{sandro.radovanovic@gmail.com}
}
\keyword{datasets}
