% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whibo_clustering.R
\name{whibo_cluster-class}
\alias{whibo_cluster-class}
\title{As S4 class to represent WhiBo Cluster model}
\description{
As S4 class to represent WhiBo Cluster model
}
\section{Slots}{

\describe{
\item{\code{whibo_cluster}}{Whibo Clustering object - list of objects for White-Box Clustering}
}}

\author{
Sandro Radovanovic
}
